test: retarget cancelClassName coverage to the code this PR changes
overlays.tsx has zero diff against master -- PlaceholderOverlay's
cancelClassName prop and its forwarding to CancelButton already
existed before this PR, so testing it directly proved nothing about
this change. Test MediaPlaceholder's forwarding and
AnimatedImagePreview's top-10 override instead, since those are what
actually changed. Verified by temporarily reverting the top-10 line
and confirming the new test fails.
